Move GetExt to core
